Fixes from PR comments
