+++ amundsen_application/static/js/components/DashboardPage/constants.ts
@@ -1,2 +1,3 @@
 export const TABLES_PER_PAGE = 10;
 export const DASHBOARD_SOURCE = "dashboard_page";
+export const DASHBOARD_OWNER_SOURCE = "dashboard_page_owner"
