+++ amundsen_application/static/js/ducks/dashboard/api/v0.ts
@@ -8,13 +8,13 @@ export type GetDashboardAPI = {
   dashboard: Dashboard;
 }
 
-const DASHBOARD_BASE = '/api/dashboard/v0';
+const DASHBOARD_BASE = '/api/metadata/v0';
 export const DASHBOARD_PREVIEW_BASE = '/api/dashboard_preview/v0/dashboard'
 
 export function getDashboard(uri: string) {
   return axios.get(`${DASHBOARD_BASE}/dashboard?uri=${uri}`)
   .then((response: AxiosResponse<GetDashboardAPI>) => {
-    return response.data;
+    return response.data.dashboard;
   });
 }
